Lab1 fixes

+++ Domain/Interfaces/IMovable.cs
@@ -2,6 +2,6 @@ namespace Spaceship2025.Domain.Interfaces;
 
 public interface IMovable
 {
-    (int X, int Y) Position { get; set; }
-    (int dX, int dY) Velocity { get;}
+    (double X, double Y) Position { get; set; }
+    (double dX, double dY) Velocity { get;}
 }
